do_bhai_banayenge_games/requirements.txt
Flask==3.1.1
pymongo==4.13.2
python-dotenv==1.1.1
